Support mocking env in integration tests
