+++ library/src/integrationTest/kotlin/com/gabrielfeo/gradle/enterprise/api/internal/KeychainIntegrationTest.kt
@@ -1,16 +1,16 @@
 package com.gabrielfeo.gradle.enterprise.api.internal
 
 import com.gabrielfeo.gradle.enterprise.api.internal.keychain
-import kotlin.test.Test
-import kotlin.test.assertFalse
-import kotlin.test.assertIs
+import org.junit.jupiter.api.Assertions.*
+import org.junit.jupiter.api.Test
 
 internal class KeychainIntegrationTest {
 
     @Test
     fun getApiToken() {
         val result = keychain.get("gradle-enterprise-api-token")
-        assertIs<KeychainResult.Success>(result)
-        assertFalse(result.token.isNullOrBlank(), "Keychain returned null or blank")
+        assertInstanceOf(KeychainResult.Success::class.java, result)
+        val success = result as KeychainResult.Success
+        assertFalse(success.token.isNullOrBlank(), "Keychain returned null or blank")
     }
 }
